+++ ingestion/src/metadata/examples/workflows/superset.yaml
@@ -1,16 +1,26 @@
 source:
   type: superset
-  serviceName: local_superset
+  serviceName: local_superset_12
   serviceConnection:
     config:
-      hostPort: http://localhost:8080
-      username: admin
-      password: admin
+      hostPort: http://localhost:8088
+      connection:
+        type: Postgres
+        username: superset
+        password: superset
+        hostPort: localhost:5432
+        database: superset
+        # username: admin
+        # password: admin
+        # provider: db
       type: Superset
   sourceConfig:
     config:
       chartFilterPattern: {}
       dashboardFilterPattern: {}
+      dbServiceNames:
+      - OMD
+      type: DashboardMetadata
 sink:
   type: metadata-rest
   config: {}
