+++ openmetadata-spec/src/main/resources/json/schema/entity/services/connections/dashboard/supersetConnection.json
@@ -11,6 +11,30 @@
       "type": "string",
       "enum": ["Superset"],
       "default": "Superset"
+    },
+    "apiConnection":{
+      "title": "API Connection",
+      "type": "object",
+      "properties": {
+        "provider": {
+          "title": "Provider",
+          "description": "Authentication provider for the Superset service. For basic user/password authentication, the default value `db` can be used. This parameter is used internally to connect to Superset's REST API.",
+          "type": "string",
+          "default": "db"
+        },
+        "username": {
+          "title": "Username",
+          "description": "Username for Superset.",
+          "type": "string"
+        },
+        "password": {
+          "title": "Password",
+          "description": "Password for Superset.",
+          "type": "string",
+          "format": "password"
+        }
+      },
+      "required": ["username","password","provider"]
     }
   },
   "properties": {
@@ -28,22 +52,20 @@
       "format": "uri",
       "default": "http://localhost:8088"
     },
-    "username": {
-      "title": "Username",
-      "description": "Username for Superset.",
-      "type": "string"
-    },
-    "password": {
-      "title": "Password",
-      "description": "Password for Superset.",
-      "type": "string",
-      "format": "password"
-    },
-    "provider": {
-      "title": "Provider",
-      "description": "Authentication provider for the Superset service. For basic user/password authentication, the default value `db` can be used. This parameter is used internally to connect to Superset's REST API.",
-      "type": "string",
-      "default": "db"
+    "connection":{
+      "title": "Superset Connection",
+      "description": "Choose between API or database connection fetch metadata from superset.",
+      "oneOf": [
+        {
+          "$ref": "#/definitions/apiConnection"
+        },
+        {
+          "$ref": "../database/postgresConnection.json"
+        },
+        {
+          "$ref": "../database/mysqlConnection.json"
+        }
+      ]
     },
     "connectionOptions": {
       "title": "Connection Options",
@@ -56,5 +78,5 @@
     }
   },
   "additionalProperties": false,
-  "required": ["hostPort", "username"]
+  "required": ["hostPort","connection"]
 }
